+++ src/core/config/npo/queries.rs
@@ -1,75 +1,37 @@
-use std::collections::{HashMap, HashSet};
 use std::fmt::{Display, Formatter};
 
-use derive_getters::Getters;
-
-use crate::core::config::npo::{FieldName, TypeName};
+use super::{chunk::Chunk, FieldName};
 
 ///
 /// Represents a list of query paths that can issue a N + 1 query
-#[derive(Default, Debug, PartialEq, Getters)]
-pub struct Queries<'a> {
-    map: HashMap<TypeName<'a>, HashSet<(FieldName<'a>, TypeName<'a>)>>,
-    root: &'a str,
-}
+#[derive(Default, Debug, PartialEq)]
+pub struct Queries<'a>(Vec<Vec<&'a str>>);
 
-impl<'a> Queries<'a> {
-    pub fn new(
-        map: HashMap<TypeName<'a>, HashSet<(FieldName<'a>, TypeName<'a>)>>,
-        root: &'a str,
-    ) -> Self {
-        Self { map, root }
+impl Queries<'_> {
+    pub fn size(&self) -> usize {
+        self.0.len()
     }
-
-    ///
-    /// Returns the query paths that can issue a N + 1 query
-    pub fn as_path(&self) -> Vec<Vec<FieldName<'a>>> {
-        let mut result = Vec::new();
-        let mut visited = HashSet::new();
-
-        #[allow(clippy::too_many_arguments)]
-        fn dfs<'a>(
-            map: &HashMap<TypeName<'a>, HashSet<(FieldName<'a>, TypeName<'a>)>>,
-            ty: TypeName<'a>,
-            path: Vec<(&'a str, (&'a str, &'a str))>,
-            result: &mut Vec<Vec<(&'a str, (&'a str, &'a str))>>,
-            visited: &mut HashSet<(TypeName<'a>, FieldName<'a>)>,
-        ) {
-            if let Some(fields) = map.get(&ty) {
-                for (field_name, ty_of) in fields {
-                    let mut new_path = path.clone();
-                    new_path.push((ty.0, (field_name.as_str(), ty_of.0)));
-                    if !visited.contains(&(ty, *field_name)) {
-                        visited.insert((ty, *field_name));
-                        dfs(map, *ty_of, new_path, result, visited);
-                        visited.remove(&(ty, *field_name));
-                    }
-                }
-            } else {
-                result.push(path);
-            }
-        }
-
-        dfs(
-            &self.map,
-            TypeName(self.root),
-            Vec::new(),
-            &mut result,
-            &mut visited,
-        );
-
-        result
-            .into_iter()
-            .map(|v| v.into_iter().map(|(_, (f, _))| FieldName::new(f)).collect())
-            .collect()
+    pub fn from_chunk<'a>(chunk: Chunk<Chunk<FieldName<'a>>>) -> Queries<'a> {
+        Queries(
+            chunk
+                .as_vec()
+                .iter()
+                .map(|chunk| {
+                    chunk
+                        .as_vec()
+                        .iter()
+                        .map(|field_name| field_name.as_str())
+                        .collect()
+                })
+                .collect(),
+        )
     }
 }
 
 impl<'a> Display for Queries<'a> {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let query_paths = self.as_path();
-
-        let query_data: Vec<String> = query_paths
+        let query_data: Vec<String> = self
+            .0
             .iter()
             .map(|query_path| {
                 let mut path = "query { ".to_string();
